// server/src/domain/interfaces/use-cases/company/IGetCandidateDetailsUseCase.ts
import { SeekerProfile, Experience, Education } from '../../../entities/seeker-profile.entity';

export interface CandidateDetails {
  profile: SeekerProfile;
  user: {
    _id: string;
    name: string;
    email: string;
  };
  experiences: Experience[];
  educations: Education[];
}

export interface IGetCandidateDetailsUseCase {
  execute(candidateId: string): Promise<CandidateDetails>;
}